extract new evidence form
